Remove Sending module
